Flytt sorteringslogikk opp i context
